# feat: make `formatMessage` take in `ReactElement`