--- a/src/components/message.tsx
+++ b/src/components/message.tsx
@@ -8,14 +8,15 @@ import * as React from 'react';
 import withIntl from './injectIntl';
 import {MessageDescriptor, IntlShape} from '../types';
 const shallowEquals = require('shallow-equal/objects');
-import {formatMessage as baseFormatMessage} from '../format';
+
+import {formatMessageToParts as baseFormatMessageToParts} from '../format';
 import {
   invariantIntlContext,
   DEFAULT_INTL_CONFIG,
   createDefaultFormatters,
 } from '../utils';
 
-const defaultFormatMessage: IntlShape['formatMessage'] = (
+const defaultFormatMessageToParts: IntlShape['formatMessageToParts'] = (
   descriptor,
   values
 ) => {
@@ -25,7 +26,7 @@ const defaultFormatMessage: IntlShape['formatMessage'] = (
     );
   }
 
-  return baseFormatMessage(
+  return baseFormatMessageToParts(
     {
       ...DEFAULT_INTL_CONFIG,
       locale: 'en',
@@ -76,7 +77,7 @@ export class BaseFormattedMessage extends React.Component<Props> {
 
   render() {
     const {
-      formatMessage = defaultFormatMessage,
+      formatMessageToParts = defaultFormatMessageToParts,
       textComponent: Text = React.Fragment,
     } = this.props.intl || {};
 
@@ -89,62 +90,8 @@ export class BaseFormattedMessage extends React.Component<Props> {
       children,
     } = this.props;
 
-    let tokenDelimiter: string = '';
-    let tokenizedValues: Record<string, string> = {};
-    let elements: Record<string, React.ReactChild> = {};
-
-    let hasValues = values && Object.keys(values).length > 0;
-    if (hasValues) {
-      // Creates a token with a random UID that should not be guessable or
-      // conflict with other parts of the `message` string.
-      let uid = Math.floor(Math.random() * 0x10000000000).toString(16);
-
-      let generateToken = (() => {
-        let counter = 0;
-        return () => `ELEMENT-${uid}-${(counter += 1)}`;
-      })();
-
-      // Splitting with a delimiter to support IE8. When using a regex
-      // with a capture group IE8 does not include the capture group in
-      // the resulting array.
-      tokenDelimiter = `@__${uid}__@`;
-
-      // Iterates over the `props` to keep track of any React Element
-      // values so they can be represented by the `token` as a placeholder
-      // when the `message` is formatted. This allows the formatted
-      // message to then be broken-up into parts with references to the
-      // React Elements inserted back in.
-      Object.keys(values).forEach(name => {
-        let value = values[name];
-
-        if (React.isValidElement(value)) {
-          let token = generateToken();
-          tokenizedValues[name] = tokenDelimiter + token + tokenDelimiter;
-          elements[token] = value;
-        } else {
-          tokenizedValues[name] = value;
-        }
-      });
-    }
-
-    let descriptor = {id, description, defaultMessage};
-    let formattedMessage = formatMessage(descriptor, tokenizedValues || values);
-
-    let nodes: React.ReactNodeArray;
-
-    let hasElements = elements && Object.keys(elements).length > 0;
-    if (hasElements) {
-      // Split the message into parts so the React Element values captured
-      // above can be inserted back into the rendered message. This
-      // approach allows messages to render with React Elements while
-      // keeping React's virtual diffing working properly.
-      nodes = formattedMessage
-        .split(tokenDelimiter)
-        .filter(Boolean)
-        .map(part => elements[part] || part);
-    } else {
-      nodes = [formattedMessage];
-    }
+    const descriptor = {id, description, defaultMessage};
+    const nodes: React.ReactNodeArray = formatMessageToParts(descriptor, values);
 
     if (typeof children === 'function') {
       return children(...nodes);