--- a/src/components/provider.tsx
+++ b/src/components/provider.tsx
@@ -98,7 +98,6 @@ export function getBoundFormatFns(
       config,
       formatterFns
     ),
-    formatMessageToParts: formatters.formatMessageToParts.bind(undefined, config, formatterFns),
     formatHTMLMessage: formatters.formatHTMLMessage.bind(
       undefined,
       config,