… docs/API.md
… docs/Upgrade-Guide.md
… package.json