@redonkulus 2019-07-12
What’s the use case for this?